pyproject.toml
[tool.poetry]
name = "apple-stock-rl-trader"
version = "0.1.0"
description = ""
authors = ["Qamil-Mirza <qamilmirza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.3"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
scikit-learn = "^1.5.2"
gymnasium = "^1.0.0"
stable-baselines3 = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
